The action type is a string that defines what should happen. ADD_COLOR is the action that will add a new color to our list of colors in the application state. It is pretty easy to make typos when creating actions using strings: <code> { type: "ADD_COOLOR" } </code> This typo would cause a bug in our application. This type of error usually does not trigger any warnings; you simply will not see the expected change of your state data. If you make these errors, they can be tough to find. This is where constants can save you: <code> import C from "./constants"
{ type: C.ADD_COLOR }
</code>
This specifies the same action, but with a JavaScript constant instead of a string. A typo in a JavaScript variable will cause the browser to throw an error. Defining actions as constants also lets you tap into the benefits of IntelliSense and code completion in your IDE. When you type the first letter or two of a constant, the IDE will autocomplete it for you. Using constants is not required, but it is not a bad idea to get into the habit of incorporating them.