Each reducer is designed to handle only the actions necessary to update its part of the state tree. The color reducer will handle only actions that require a new or changed color object: ADD_COLOR and RATE_COLOR. The colors reducer will focus on those actions necessary for managing the colors array: ADD_COLOR, REMOVE_COLOR, RATE_COLOR. Finally, the sort reducer will handle the SORT_COLORS action.
Each reducer is composed or combined into a single reducer function that will use the store. The colors reducer is composed with the color reducer to manage individual colors within the array. The sort reducer will then be combined with the colors reducer to create a single reducer function. This can update our entire state tree and handle any action sent to it.

Both the colors and color reducers will handle ADD_COLOR and RATE_COLOR. But remember, each reducer focuses on a specific part of the state tree. RATE_COLOR in the color reducer will handle the task of changing an individual color’s rating value; RATE_COLOR in the colors reducer will focus on locating the color that needs to be rated in the array. ADD_COLOR in the color reducer will result in a new color object with the correct properties; ADD_COLOR in the colors reducer will return an array that has an additional color object. They are meant to work together. Each reducer focuses on what a specific action means for its branch in the state tree.

REDUCER COMPOSITION IS NOT REQUIRED, JUST RECOMMENDED
Redux does not require that we create smaller, more focused reducers and compose them into a single reducer. We could create one reducer function to handle every action in our app. In doing so, we would lose the benefits of modularity and functional programming.

Here are the actions for the color reducer:

ADD_COLOR
Returns a new color object constructed from the action’s payload data.

RATE_COLOR
Returns a new color object with the desired rating. The ES7 object spread operator allows us to assign the value of the current state to a new object.

Reducers should always return something. If for some reason this reducer is invoked with an unrecognized action, we will return the current state: the default case.