The Colors Reducer The color reducer is designed to manage leaves on the colors branch of our state tree. The colors reducer will be used to manage the entire colors branch. The colors reducer will handle any actions for adding, rating, and removing colors.
ADD_COLOR
Creates a new array by concatenating all of the values of the existing state array with a new color object. The new color is created by passing a blank state object and the action to the color reducer.
RATE_COLOR
Returns a new array of colors with the desired color rated. The colors reducer locates the color to be rated within the current state array. It then uses the color reducer to obtain the newly rated color object and replaces it in the array.
REMOVE_COLOR
Creates a new array by filtering out the desired color to remove.
The colors reducer is concerned with the array of colors. It uses the color reducer to focus on the individual color objects.

TREAT STATE AS AN IMMUTABLE OBJECT
In all of these reducers, we need to treat state as an immutable object. Although it may be tempting to use state.push({}) or state[index].rating, we should resist the urge to do so.

NO SIDE EFFECTS IN REDUCERS
Reducers should be predictable. They are used to simply manage the state data. In the previous example, notice that the timestamp and IDs are generated prior to sending the action to the reducer. Generating random data, calling APIs, and other asynchronous processes should be handled outside of reducers. Avoiding state mutations and side effects is always recommended.