The sort reducer is an entire function designed to manage one string variable in our state:
The sort reducer is used to change the sort state variable. It sets the sort state to the value of the action’s sortBy field (if this is not a state provided, it will return SORTED_BY_DATE):

To recap, state updates are handled by reducers. Reducers are pure functions that take in state as the first argument and an action as the second argument. Reducers do not cause side effects and should treat their arguments as immutable data. In Redux, modularity is achieved through reducers. Eventually, reducers are combined into a single reducer, a function that can update the entire state tree.

In this section, we saw how reducers can be composed. We saw how the colors reducer uses the color reducer to assist in color management.