The Store In Redux, the store is what holds the application’s state data and handles all state updates.5 While the Flux design pattern allows for many stores that each focus on a specific set of data, Redux only has one store.
The store handles state updates by passing the current state and action through a single reducer. We will create this single reducer by combining and composing all of our reducers.

In order to create a single reducer tree, we must combine the colors and sort reducers. Redux has a function for doing just that, combineReducers, which combines all of the reducers into a single reducer. These reducers are used to build your state tree. The names of the fields match the names of the reducers that are passed in.

A store can also be created with initial data. Invoking the colors reducer without state returns an empty array:

The only way to change the state of your application is by dispatching actions through the store. The store has a dispatch method that is ready to take actions as an argument. When you dispatch an action through the store, the action is sent through the reducers and the state is updated: