Let’s refactor the AddColorForm component to retrieve the store and dispatch the ADD_COLOR action directly. The component can retrieve the store and dispatch ADD_COLOR actions directly. In order to use the store, we must define contextTypes on the AddColorForm instance. This is where we tell React which context variables this component will use. This is a required step. Without it, the store cannot be retrieved from the context. A component class, and can access context via this.context.
The context object is passed to stateless functional components as the second argument, after props. We can use object destructuring to obtain the store from this object directly in the arguments. The same rules apply that do for stateless functional components. contextTypes must be defined on the instance.

Retrieving the store from the context is a nice way to reduce your boilerplate, but this is not something that is required for every application. Dan Abramov, the creator of Redux, even suggests that these patterns do not need to be religiously followed:

Separating the container and presentational components is often a good idea, but you shouldn’t take it as dogma. Only do this when it truly reduces the complexity of your codebase.2