When the application starts, instead of rendering a single App component, we will render the HashRouter component.
The HashRouter component is rendered as the root component for our application. Each route can be defined within the HashRouter using the Route component. These routes tell the router which component to render when the window’s location changes. Each Route component has path and component properties. When the browser’s location matches the path, the component will be displayed. When the location is /, the router will render the Home component. When the location is /products, the router will render the Products component.The first route, the one that displays the Home component, has an exact property. This means that the Home component will only be displayed when the location exactly matches the root /.

At this point, we can run the app and physically type the routes into the browser’s location bar to watch the content change. For example, type http://localhost:3000/#/about into the location bar and watch the About component render.